add User.CanonicalMe field
This new field is needed for identifying users
that authenticate via the IndieAuth protocol.

It may have been possible to re-use the HTMLURL field for that,
but it was deemed better to add a very explicit dedicated field
for something as critical as identification and authentication.

Improve documentation and update internal notes.

Updates https://github.com/shurcooL/home/issues/34.